+++ database/migrations/2020_02_19_115638_create_economic_activity_settlement_table.php
@@ -15,7 +15,7 @@ class CreateEconomicActivitySettlementTable extends Migration
     {
         Schema::create('economic_activity_settlement', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->float('amount');
+            $table->decimal('amount', 15, 2);
             $table->unsignedBigInteger('economic_activity_id');
             $table->unsignedBigInteger('settlement_id');
             $table->foreign('economic_activity_id')->references('id')
